+++ .gitignore
@@ -1,2 +1,3 @@
 deno.lock
 examples/*.tscn
+scenes/
