+++ lib/main.ts
@@ -1,0 +1,4 @@
+export * from "./core/nodes/collision-shape-2d.tsx";
+export * from "./core/nodes/physics/character-body-2d.tsx";
+export * from "./core/resources/shapes/rectangle-shape-2d.ts";
+export * from "./core/godact.ts";
