+++ main.tsx
@@ -1,24 +1,4 @@
 import Player from "./examples/player.tsx";
-import { reconciler } from "./lib/internal/reconciler.ts";
-import { createRenderer } from "./lib/internal/renderers/renderer.ts";
-const container = createRenderer();
+import { createGodactScene } from "./lib/main.ts";
 
-const root = reconciler.createContainer(
-  container,
-  0,
-  null,
-  false,
-  null,
-  "godact",
-  (error: Error) => console.error(error),
-  null,
-);
-
-if (root !== null) {
-  reconciler.updateContainer(<Player />, root, null);
-
-  Deno.writeFileSync(
-    "./examples/player.tscn",
-    new TextEncoder().encode(container.compileScript()),
-  );
-}
+createGodactScene(<Player />, "./scenes/player.tscn");
