+++ main_test.tsx
@@ -15,9 +15,10 @@ Deno.test("Test render", () => {
     null,
     "godact",
     (error: Error) => console.error(error),
-    null,
+    null
   );
 
   reconciler.updateContainer(<Player />, root, null);
+
   assertEquals(container.compileScript(), expectedScript);
 });
