+++ tests/player.tsx
@@ -1,6 +1,8 @@
-import { CharacterBody2D } from "../lib/core/nodes/physics/character-body-2d.tsx";
-import { CollisionShape2D } from "../lib/core/nodes/collision-shape-2d.tsx";
-import { createRectangleShape2D } from "../lib/core/resources/shapes/rectangle-shape-2d.ts";
+import {
+  CharacterBody2D,
+  CollisionShape2D,
+  createRectangleShape2D,
+} from "../lib/main.ts";
 
 export default function Player() {
   return (
